test: Add C++ guards
When using gTest with C++ these files included errors with missing pre
declaration and extern guards.

BUG=b:323908896
TEST=zmake build --all
LOW_COVERAGE_REASON=Coverage is added in chromium:5273875, b:319290529

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5273864
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>

+++ zephyr/shim/include/shimmed_task_id.h
@@ -276,7 +276,7 @@ enum {
 };
 
 #define EXTRA_TASK_ID(name) \
-	TASK_ID_##name = (TASK_ID_COUNT + EXTRA_TASK_##name),
+	TASK_ID_##name = ((int)TASK_ID_COUNT + (int)EXTRA_TASK_##name),
 enum { CROS_EC_EXTRA_TASKS(EXTRA_TASK_ID) };
 
 #endif /* __CROS_EC_SHIMMED_TASK_ID_H */
